+++ src/core/request/engines/fetch/index.ts
@@ -152,6 +152,7 @@ const request: RequestEngine = (params) => {
 				important: p.important,
 
 				okStatuses: p.okStatuses,
+				noContentStatuses: p.noContentStatuses,
 				status: res.status,
 				statusText: res.statusText,